Read integers, better this time

+++ src/Ji/Ast.fs
@@ -1,3 +1,3 @@
 module Ji.Ast
 
-type Expr = ExprInt of int
+type Expr = ExprInt of uint64
